requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
datasets>=2.14.0
pyyaml>=6.0
matplotlib>=3.7.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
